Coupled Model Components

CouplerMachine provides a wrapper for model components so that they may connect to the coupler. The CplModel struct packages a component model with the information needed for the CplSolver to evolve it.

CouplerMachine.CplModel — Type

CplModel(; 
    grid,
    equations,
    nsteps::Int,
    boundary_z = 0.0,
    dt = 1.0,
    timestepper = LSRK54CarpenterKennedy,
    NFfirstorder = RusanovNumericalFlux(),
    NFsecondorder = CentralNumericalFluxSecondOrder(),
    NFgradient = CentralNumericalFluxGradient(),
    overint_params = nothing,
    callbacks = (),
)

Builds an instance of a coupler test model. This is currently a toy model used for testing and designing coupling machinery. In a full-blown coupled experiment this model would be replaced by a full component model that is wrapped to format it for coupler compatibility.

grid the spectral element grid used by this model.

equations the Balance Law used by this model.

nsteps number of component steps to run during each coupling step.

boundary_z height above or below the air-sea interface of the coupled boundary.

dt component timestep to use on each component step.

timestepper the ODE solver used to advance the system.

NFfirstorder numerical flux to use for first order terms.

NFsecondorder numerical flux to use for second order terms.

NFgradient numerical flux to use for gradient terms.

overint_params overintegration parameters.

callbacks component-specific callbacks executed at inner time steps.

Each returned model instance is independent and has its own grid, balance law, time stepper and other attributes. For now the code keeps some of these things the same for initial testing, including component timestepper and initial time (both of which need tweaking to use for real setups).

A real model might have many more flags and/or may wrap the component creation very differently. Any component should allow itself to set a number of timesteps to execute with a certain timestep to synchronize with the coupling time scale.