Coupler State

The coupler provides a space to store information being passed between coupled model components at their boundaries. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, filtering, etc.

The ClimaCoupler defines a type CouplerState for a container variable that holds information about the field boundary values that are being used to couple components. Components can use a coupler_put! operation to export a set of field values to a CouplerState variable. A coupler_get operation is used to retrieve a set of field values from a CplState variable. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, or filtering.

Coupler Object API

ClimaCoupler.CouplerState — Type

mutable struct CouplerState{FT, CF, RO}

Type for holding coupled fields.

This is the namespace through which coupled components communicate. Its role is to provide a level of indirection so that components remain modular and so that any data communication, interpolation, reindexing/unit conversions and filtering etc... can be embeded in the intermdediate coupling layer.

A field is exported by one component and imported by one or more other components.

Fields:

coupled_fields

A dictionary of fields added to the coupler

remap_operators

A dictionary of remap operators between components

Δt_coupled

The coupled timestep size

Contructors:

CouplerState(coupled_fields, remap_operators, Δt_coupled)

defined at /home/runner/work/ClimaCoupler.jl/ClimaCoupler.jl/src/CouplerState/coupler_state.jl:29.

CouplerState(Δt_coupled)

defined at /home/runner/work/ClimaCoupler.jl/ClimaCoupler.jl/src/CouplerState/coupler_state.jl:42.

ClimaCoupler.coupler_add_field! — Function

coupler_add_field!(
    coupler,
    fieldname,
    fieldvalue;
    write_sim,
    metadata
)

Add a field to the coupler that is accessible with key fieldname.

Arguments

coupler: coupler object the field is added to.

fieldname: key to access the field in the coupler.

fieldvalue: data array of field values.

write_sim: the simulation can write to this field in the coupler.

metadata: a catch-all storage for any metadata associated with the field.

ClimaCoupler.coupler_add_map! — Function

coupler_add_map!(coupler, map_name, map)

Add a map to the coupler that is accessible with key mapname.

Arguments

coupler: coupler object the field is added to.

mapname: key to access the map in the coupler's map list.

map: a remap operator.

ClimaCoupler.coupler_push! — Function

coupler_push!(coupler, sim)

Update coupler with fields retrieved from the coupler.

coupler_push! is an adapter function to be implemented for each model component using the coupler. It should send coupling fields via coupler_put! calls and perform any operations on these fields to prepare them for the coupler.

ClimaCoupler.coupler_pull! — Function

coupler_pull!(sim, coupler)

Update model with fields retrieved from the coupler.

coupler_pull! is an adapter function to be implemented for each model component using the coupler. It should get coupling fields via coupler_get calls and perform any operations on these fields to prepare them for use in the component model.

ClimaCoupler.coupler_put! — Function

coupler_put!(coupler, fieldname, fieldvalue, source_sim)

Sets coupler field fieldname to fieldvalue.

Arguments

coupler: coupler object the field is added to.

fieldname: key to access the field in the coupler.

fieldvalue: the field being stored.

source_sim: the simulation that fieldvalue belongs to.

ClimaCoupler.coupler_get — Function

coupler_get(coupler, fieldname)

Retrieve coupler field fieldname without remapping.

Arguments

coupler: coupler object being accessed.

fieldname: key to access the field in the coupler.

coupler_get(coupler, fieldname, target_sim)

Retrieve coupler field fieldname and remap to target_sim's boundary space.

Arguments

coupler: coupler object being accessed.

fieldname: key to access the field in the coupler.

target_sim: the simulation that target_field belongs to.

ClimaCoupler.coupler_get! — Function

coupler_get!(target_field, coupler, fieldname, target_sim)

Retrieve coupler field fieldname, remap and store in target_field.

Arguments

target_field: the field to be updated by the coupler.

coupler: coupler object being accessed.

fieldname: key to access the field in the coupler.

target_sim: the simulation that target_field belongs to.