PostProcessor

This module contains functions for postprocessing model data that was saved during the simulation by ClimaCoupler.Diagnostics. This module is used for offline regridding, slicing and spatial averages. It can also handle data from other sources (e.g., NCEP reanalysis).

Diagnostics API

ClimaCoupler.PostProcessor.postprocess — Function

postprocess(
    name::Symbol,
    raw_data::Union{Fields.Field, Array},
    p_methods::Tuple;
    lev_slice = 1,
    datafile_latlon = nothing,
    nlat = 90,
    nlon = 180,
)

Coordinates regridding, averaging or slicing of variable name corresponding to raw_data. Postprocessing methods are specified in p_methods. raw_data is assumed to be a Field (dimensions corresponding to the model's CGLL grid), a 2D Array with [longitude, latitude] or a 3D Array [longitude, latitude, level].

Arguments:

name: [Symbol] variable name

raw_data: [Union{Fields.Field, Array}] variable data

p_methods: [Tuple] postproessing methods (:regrid, :horizontal_slice, :zonal_mean)

lev_slice: [Int] level index along which the :horizontal_slice is applied

datafile_latlon: [String] name of the regrid file

nlat: [Int] number of latitudes of the regridded array

nlon: [Symbol] number of longitudes of the regridded array

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.PostProcessedData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.ZLatLonData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.ZLatData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.LatLonData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.LatData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.PostProcessor.RawData. Check Documenter's build log for details.

ClimaCoupler.PostProcessor.DataPackage — Type

DataPackage(tag::PostProcessedData, name::Symbol, data::Union{Array, Field}; coords = coords)

A container for storing the tyoe, name, data and coordinates of a variable.