Regridder

This module contains functions to regrid information between spaces. Many of the functions used in this module call TempestRemap functions via ClimaCoreTempestRemap wrappers.

Information about the TempestRemap library can be found here. Multiple remapping approaches from TempestRemap have been tested with our implementation, and information about them is located here.

Regridder API

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.write_to_hdf5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.read_from_hdf5. Check Documenter's build log for details.

ClimaCoupler.Regridder.dummmy_remap! — Function

dummmy_remap!(target, source)

Simple stand-in function for remapping. For AMIP we don't need regridding of surface model CC.Fields. When we do, we re-introduce the ClimaCoreTempestRemap remapping functions.

Arguments

target: [CC.Fields.Field] destination of remapping.

source: [CC.Fields.Field] source of remapping.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.remap_field_cgll_to_rll. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.land_fraction. Check Documenter's build log for details.

ClimaCoupler.Regridder.update_surface_fractions! — Function

update_surface_fractions!(cs::Interfacer.CoupledSimulation)

Updates dynamically changing area fractions. Maintains the invariant that the sum of area fractions is 1 at all points.

Arguments

cs: [Interfacer.CoupledSimulation] containing area fraction information.

ClimaCoupler.Regridder.combine_surfaces! — Function

combine_surfaces!(combined_field::CC.Fields.Field, sims, field_name::Val)

Sums the fields, specified by field_name, weighted by the respective area fractions of all surface simulations. THe result is saved in combined_field.

Arguments

combined_field: [CC.Fields.Field] output object containing weighted values.

sims: [NamedTuple] containing simulations .

field_name: [Val] containing the name Symbol of the field t be extracted by the Interfacer.get_field functions.

Example

combine_surfaces!(temp_field, cs.model_sims, Val(:surface_temperature))

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.combine_surfaces_from_sol!. Check Documenter's build log for details.

Regridder Internal Functions

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.reshape_cgll_sparse_to_field!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.hdwrite_regridfile_rll_to_cgll. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.write_datafile_cc. Check Documenter's build log for details.

ClimaCoupler.Regridder.binary_mask — Function

binary_mask(var, threshold)

Converts a number var to 1, if var is greater or equal than a given threshold value, or 0 otherwise, keeping the same type.

Arguments

var: [FT] value to be converted.

threshold: [FT] cutoff value for conversions.

binary_mask(var)

Converts a number var to 1, if var is greater or equal than eps(FT), or 0 otherwise, keeping the same type.

Arguments

var: [FT] value to be converted.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.read_remapped_field. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.get_coords. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Regridder.get_time. Check Documenter's build log for details.