TestHelper

This module defines helper functions, objects, and constants to be used by various files in the test folder.

TestHelper API

ClimaCoupler.TestHelper.create_space — Function

create_space(FT; comms_ctx = ClimaComms.SingletonCommsContext(), 
    R = FT(6371e3), ne = 4, polynomial_degree = 3, nz = 1)

Initialize a space on a sphere with the given parameters. Used for debugging and testing.

Arguments

FT: [DataType] Float type

comms_ctx: [ClimaComms.AbstractCommsContext] context used for this operation.

R: [FT] radius of the sphere underlying space.

ne: [Integer] number of elements used in the space's mesh.

polynomial_degree: [Integer] degree of the polynomial used to represent the space (number of GLL nodes - 1).

nz: [Integer] number of vertical elements

ClimaCoupler.TestHelper.gen_ncdata — Function

gen_ncdata(FT, path, varname, val)

Create an NCDataset with lat/lon dimensions containing the value val for the variable varname, and store it at path.

Arguments

FT: [DataType] Float type.

path: [String] location to store output datafile.

varname: [Symbol] variable name.

val: [FT] value to store as varname at all indices.