Coupled Timestepping

CouplerMachine currently uses a sequential timestepping framework in which one component steps forward before passing its updated state to another. An outer "coupled" timestep determines when component models synchronize and coordinate with the coupler. Within this coupled timestep, components take an integer number of substeps, and evolve independently from each other.

CplSolver extends the ODE solver API of ClimateMachine.ODESolvers.

CouplerMachine.CplSolver — Type

CplSolver(; component_list, coupler::CplState, coupling_dt, t0)

A time stepping like object for advancing a coupled system made up of a pre-defined set of named components specified in component_list. Each component is a balance law, discretization and timestepper collection. The coupler will step them forward by a nsteps substeps to advance the coupled system by a simulated time coupling_dt.

Components interact with the coupler during pre_step() and post_step() functions. During a pre_step(), a component may get fields from the coupler name space. A component may put fields into the coupler name space during the post_step() for later use by other components. The CplSolver abstraction controls

the outer time stepping sequencing of components

the execution of actions mapping exports from one or more components to imports of other components through an intermediary coupler name space.