Utilities

This module contains functions, objects, and constants used by various modules in the coupler.

Utilities API

ClimaCoupler.Utilities.swap_space! — Function

swap_space!(space_out::CC.Spaces.AbstractSpace, field_in::CC.Fields.Field)

Remap the values of a field onto a new space.

Arguments

space_out: [CC.Spaces.AbstractSpace] The axes of the space we want to remap onto

field_in: [CC.Fields.Field] to be remapped to new space.

ClimaCoupler.Utilities.get_comms_context — Function

get_comms_context(config_dict)

Sets up the appropriate ClimaComms context for the device the model is to be run on

Arguments

config_dict: dictionary containing a "device" flag whcih decides which device context is needed

ClimaCoupler.Utilities.get_device — Function

get_device(config_dict)

Returns the device on which the model is being run

Arguments

config_dict: dictionary containing a "device" flag which decides which device to run on