Checkpointer

This module contains general functions for logging the model states and restarting simulations. The Checkpointer uses ClimaCore.InputOutput infrastructure, which allows it to handle arbitrarily distributed logging and restart setups.

Checkpointer API

ClimaCoupler.Checkpointer.get_model_state_vector — Function

get_model_state_vector(sim::Interfacer.ComponentModelSimulation)

Returns the model state of a simulation as a ClimaCore.FieldVector. This is a template function that should be implemented for each component model.

ClimaCoupler.Checkpointer.restart_model_state! — Function

restart_model_state!(sim::Interfacer.ComponentModelSimulation, comms_ctx::ClimaComms.AbstractCommsContext, t::Int; input_dir = "input")

Sets the model state of a simulation from a HDF5 file from a given time, t (in seconds).

ClimaCoupler.Checkpointer.checkpoint_model_state — Function

checkpoint_model_state(sim::Interfacer.ComponentModelSimulation, comms_ctx::ClimaComms.AbstractCommsContext, t::Int; output_dir = "output")

Checkpoints the model state of a simulation to a HDF5 file at a given time, t (in seconds).