Coupler State

The coupler provides a space to store information being passed between coupled model components at their boundaries. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, filtering, etc.

The CouplerMachine defines a type CplState for a container variable that holds information about the field boundary values that are being used to couple components. Components can use a coupler_put! operation to export a set of field values to a CplState variable. A coupler_get operation is used to retrieve a set of field values from a CplState variable. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, or filtering.

Coupler Object API

CouplerMachine.CplState — Type

CplState()

Type for holding coupler "state". This is the namespace through which coupled components communicate. Its role is to provide a level of indirection so that components remain modular and so that any data communication, interpolation, reindexing/unit conversions and filtering etc... can be embeded in the intermdediate coupling layer.

A field is exported by one component and imported by one or more other components.

CouplerMachine.coupler_register! — Function

Coupling.coupler_register!(
        coupler::CplState,
        fieldname::Symbol,
        fieldvalue,
        grid,
        datetime::DateTime,
        units::Unitful.Units = Unitful.NoUnits, 
    )

Add a field to the coupler that is accessible with key fieldname.

Arguments

coupler: coupler object the field is registered to.

fieldname: key to access the field in the coupler.

fieldvalue: data array of field values.

grid: grid the field is stored on.

datetime: time associated with the field state.

units: units associated with the field values. Dimensionless by default.

CouplerMachine.coupler_put! — Function

coupler_put!(coupler::CplState, fieldname::Symbol, fieldvalue, gridinfo, datetime::DateTime, units::Unitful.Units)

Updates coupler field fieldname with fieldvalue, the field's value at time datetime.

gridinfo and units inform the coupler of the format of the inputted data allowing conversion to match the grid and units of the coupler field.

CouplerMachine.coupler_get — Function

coupler_get(coupler::CplState, fieldname::Symbol, gridinfo, datetime::DateTime, units::Unitful.Units)

Retrieve data array corresponding to fieldname.

Returns data on the grid specified by gridinfo and in the units of units. Checks that the coupler data field is the state at time datetime.