Coupler State

The coupler provides a space to store information being passed between coupled model components at their boundaries. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, filtering, etc.

The ClimaCoupler defines a type CouplerState for a container variable that holds information about the field boundary values that are being used to couple components. Components can use a coupler_put! operation to export a set of field values to a CouplerState variable. A coupler_get operation is used to retrieve a set of field values from a CplState variable. During this exchange, the coupler manages ancillary operations such as regridding, unit conversions, or filtering.

Coupler Object API

ClimaCoupler.CouplerState — Type

CouplerState()

Type for holding coupler "state". This is the namespace through which coupled components communicate. Its role is to provide a level of indirection so that components remain modular and so that any data communication, interpolation, reindexing/unit conversions and filtering etc... can be embeded in the intermdediate coupling layer.

A field is exported by one component and imported by one or more other components.

ClimaCoupler.coupler_add_field! — Function

coupler_add_field!(
        coupler::CouplerState,
        fieldname::Symbol,
        fieldvalue,
    )

Add a field to the coupler that is accessible with key fieldname.

Arguments

coupler: coupler object the field is added to.

fieldname: key to access the field in the coupler.

fieldvalue: data array of field values.

ClimaCoupler.coupler_put! — Function

coupler_put!(coupler::CouplerState, fieldname::Symbol, fieldvalue)

Sets coupler field fieldname to fieldvalue.

ClimaCoupler.coupler_get — Function

coupler_get(coupler::CouplerState, fieldname::Symbol [, target_sim::AbstractSimulation])

Retrieve data array corresponding to fieldname.

If a target_sim is passed, the field is remapped to that simulation's boundary space.