Diagnostics

This module contains functions for defining, gathering and outputting model diagnostics from the Coupler.

Note that ClimaCoupler.Diagnostics is deployed online (i.e., as the model runs), working with cached model data (usually) on the model grid. This does not include offline post-processing (i.e., manipulating saved model output after the model is run, such as regridding data to the latitude-longitude grid). See ClimaCoupler.PostProcessor for offline model data treatment.

Diagnostics API

Missing docstring.

Missing docstring for ClimaCoupler.Diagnostics.AbstractOutputGroup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Diagnostics.DiagnosticsGroup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Diagnostics.AbstractDiagnosticsOperations. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.Diagnostics.TimeMean. Check Documenter's build log for details.

ClimaCoupler.Diagnostics.get_var — Function

get_var(cs::CoupledSimulation, x)

Defines variable extraction from the coupler simulation. User specific diagnostics should extend this function in the experiments folder.

Example:

getvar(cs, ::Val{:Tsfc}) = cs.fields.T_S

ClimaCoupler.Diagnostics.accumulate_diagnostics! — Function

accumulate_diagnostics!(cs::CoupledSimulation)

Accumulates user-defined diagnostics listed in the in the field_vector of each dg.

ClimaCoupler.Diagnostics.save_diagnostics — Function

save_diagnostics(cs::CoupledSimulation)

save_diagnostics(cs::CoupledSimulation, dg::DiagnosticsGroup, output_dir::String)

Saves all entries in dg in separate HDF5 files per variable in output_dir.

Diagnostics Internal Functions

ClimaCoupler.Diagnostics.collect_diags — Function

collect_diags(cs::CoupledSimulation, dg::DiagnosticsGroup)

Collects diagnostics in diags names.

Missing docstring.

Missing docstring for ClimaCoupler.Diagnostics.iterate_operations. Check Documenter's build log for details.

ClimaCoupler.Diagnostics.operation — Function

operation(cs::CoupledSimulation, dg::DiagnosticsGroup, new_diags::Fields.FieldVector, ::TimeMean)

Accumulates in time all entries in new_diags and saves the result in dg.field_vector, while increasing the dg.ct counter.

operation(cs::CoupledSimulation, dg::DiagnosticsGroup, new_diags::Fields.FieldVector, ::Nothing)

Accumulates in time all entries in new_diags and saves the result in dg.field_vector, while increasing the dg.ct counter.

ClimaCoupler.Diagnostics.pre_save — Function

pre_save(::TimeMean, cs::CoupledSimulation, dg::DiagnosticsGroup)

Divides the accumulated sum by 'ct' to form the mean, before saving the diagnostics.

pre_save(::Nothing, cs::CoupledSimulation, dg::DiagnosticsGroup

Collects variables and performs all specified operations before saving the snapshot diagnostics.

ClimaCoupler.Diagnostics.post_save — Function

post_save(::TimeMean, cs::CoupledSimulation, dg::DiagnosticsGroup)

Resets accumulating fields and counts after saving the diagnostics.

ClimaCoupler.Diagnostics.save_time_format — Function

save_time_format(date::Dates.DateTime, ::Monthly)

Converts the DateTime date to the conventional Unix format (seconds elapsed since 00:00:00 UTC on 1 January 1970).