FluxCalculator

This modules contains abstract types and functions to calculate surface fluxes in the coupler, or to call flux calculating functions from the component models.

Fluxes over a heterogeneous surface (e.g., from a gridpoint where atmospheric cell is overlying both land and ocean) can be handled in two different ways:

Combined fluxes (called with CombinedStateFluxes())

these are calculated by averaging the surface properties for each gridpoint (e.g., land and ocean temperatures, albedos and roughness lengths are averaged, based on their respective area fractions), so the flux is calculated only once per gridpoint of the grid where we calculate fluxes. This is computationally faster, but it makes the fluxes on surface boundaries more diffuse. Currently, we use this method for calculating radiative fluxes in the atmosphere, and turbulent fluxes in the coupler (on the atmospheric grid). The fluxes are calculated in the atmospheric (host) model's cache, which can be setup to avoid allocating coupler fields.

Partitioned fluxes (called with PartitionedStateFluxes())

these are calculated separately for each surface type. It is then the fluxes (rather than the surface states) that are combined and passed to the atmospheric model as a boundary condition. This method ensures that each surface model receives fluxes that correspond to its state properties, resulting in a more accurate model evolution. However, it is more computationally expensive, and requires more communication between the component models.

FluxCalculator API

ClimaCoupler.FluxCalculator.TurbulentFluxPartition — Type

TurbulentFluxPartition

Abstract type for flags that denote where and how to calculate tubulent fluxes.

ClimaCoupler.FluxCalculator.PartitionedStateFluxes — Type

PartitionedStateFluxes <: TurbulentFluxPartition

A flag indicating that the turbulent fluxes should be partitioned and calculated over each surface model and then combined. This is calculated on the coupler grid.

ClimaCoupler.FluxCalculator.CombinedStateFluxes — Type

CombinedStateFluxes <: TurbulentFluxPartition

A flag indicating that the turbulent fluxes (e.g. sensible and latent heat fluxes, drag and moisture fluxes) are to be calculated on the Atmos grid, and saved in Atmos cache.

ClimaCoupler.FluxCalculator.combined_turbulent_fluxes! — Function

combined_turbulent_fluxes!(model_sims, csf, turbulent_fluxes::TurbulentFluxPartition)

Calls the method(s) which calculate turbulent surface fluxes from combined surface states in coupler fields, csf.

Arguments

model_sims: [NamedTuple] containing ComponentModelSimulations.

csf: [NamedTuple] containing coupler fields.

turbulent_fluxes: [TurbulentFluxPartition] denotes a flag for turbulent flux calculation.

ClimaCoupler.FluxCalculator.atmos_turbulent_fluxes! — Function

atmos_turbulent_fluxes!(sim::Interfacer.ComponentModelSimulation, csf)

A function to calculate turbulent surface fluxes using the combined surface states. It is required that a method is defined for the given sim and that the fluxes are saved in that sim's cache. csf refers to the coupler fields.

Arguments

sim: [Interfacer.ComponentModelSimulation] object containing the component model simulation.

csf: [NamedTuple] containing coupler fields.

Example:

function atmos_turbulent_fluxes!(atmos_sim::ClimaAtmosSimulation, csf)
    atmos_sim.cache.flux .= atmos_sim.c .* (csf.T_S .- atmos_sim.temperature)
end