PostProcessor

This module contains functions for postprocessing model data that was saved during the simulation by ClimaCoupler.Diagnostics. This module is used for offline regridding, slicing and spatial averages. It can also handle data from other sources (e.g., NCEP reanalysis).

Diagnostics API

ClimaCoupler.PostProcessor.postprocess — Function

postprocess(
    name::Symbol,
    raw_data::Union{CC.Fields.Field, Array},
    p_methods::Tuple;
    lev_slice = 1,
    datafile_latlon = nothing,
    nlat = 90,
    nlon = 180,
)

Coordinates regridding, averaging or slicing of variable name corresponding to raw_data. Postprocessing methods are specified in p_methods. raw_data is assumed to be a Field (dimensions corresponding to the model's CGLL grid), a 2D Array with [longitude, latitude] or a 3D Array [longitude, latitude, level].

Arguments:

name: [Symbol] variable name

raw_data: [Union{CC.Fields.Field, Array}] variable data

p_methods: [Tuple] postproessing methods (:regrid, :horizontal_slice, :zonal_mean)

lev_slice: [Int] level index along which the :horizontal_slice is applied

datafile_latlon: [String] name of the regrid file

nlat: [Int] number of latitudes of the regridded array

nlon: [Symbol] number of longitudes of the regridded array

ClimaCoupler.PostProcessor.PostProcessedData — Type

PostProcessedData

Abstract type for postprocessed data.

ClimaCoupler.PostProcessor.ZLatLonData — Type

ZLatLonData <: PostProcessedData

Concrete type for 3D data.

ClimaCoupler.PostProcessor.ZLatData — Type

ZLatData <: PostProcessedData

Concrete type for 2D data with latitude and level.

ClimaCoupler.PostProcessor.LatLonData — Type

LatLonData <: PostProcessedData

Concrete type for 2D data with latitude and longitude.

ClimaCoupler.PostProcessor.LatData — Type

LatData <: PostProcessedData

Concrete type for 1D data with latitude.

ClimaCoupler.PostProcessor.RawData — Type

RawData <: PostProcessedData

Concrete type for raw model data.

ClimaCoupler.PostProcessor.DataPackage — Type

DataPackage(tag::PostProcessedData, name::Symbol, data::Union{Array, Field}; coords = coords)

A container for storing the tyoe, name, data and coordinates of a variable.