TimeManager

This module contains functions that handle dates and times in simulations. The functions in this module often call functions from Julia's Dates module.

TimeManager API

ClimaCoupler.TimeManager.current_date — Function

current_date(cs::CoupledSimulation, t::Int)

Return the model date at the current timestep.

Arguments

cs: [CoupledSimulation] containing info about the simulation

t: [Real] number of seconds since simulation began

ClimaCoupler.TimeManager.strdate_to_datetime — Function

strdate_to_datetime(strdate::String)

Convert from String ("YYYYMMDD") to Date format, required by the official AMIP input files.

Arguments

strdate: [String] to be converted to Date type

ClimaCoupler.TimeManager.datetime_to_strdate — Function

datetime_to_strdate(datetime::DateTime)

Convert from Date to String ("YYYYMMDD") format.

Arguments

datetime: [DateTime] object to be converted to string

ClimaCoupler.TimeManager.trigger_callback — Function

trigger_callback(cs, ::Monthly)

Returns true if the current date is equal to or exceeds the saved first of the month at time of 00:00:00.

Arguments

cs: [CoupledSimulation] containing info about the simulation

Missing docstring.

Missing docstring for ClimaCoupler.TimeManager.Monthly. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaCoupler.TimeManager.EveryTimestep. Check Documenter's build log for details.