Coupled Timestepping

CouplerMachine.CplSolver — Type

CplSolver(; component_list, coupler::CplState, coupling_dt, t0)

A time stepping like object for advancing a coupled system made up of a pre-defined set of named components specified in component_list. Each component is a balance law, discretization and timestepper collection. The coupler will step them forward by a nsteps substeps to advance the coupled system by a simulated time coupling_dt.

Components are registered with prestep() and poststep() functions. The prestep() functions get fields for use in the component from the coupler name space. The poststep() functions put fields for use by other components into te coupler name space. The CplSolver abstraction controls

the outer time stepping sequencing of components

the excution of actions mapping exports from one or more components to imports of other components through an intermediary coupler name space.

Some notes -

For now components need to include slightly wasteful "shadow" variables for accumulating boundary flux terms they compute across RK stages and across timesteps. These are defined within the component balance law. The shadown variable is a full 3d array because of the way the current infrastructure works. This can be tidied up later once design is settled.