Simulations & Timestepping

Simulations Interface

ClimaCoupler.jl organizes coupled models and their execution through the use of Simulations. A Simulation wraps a physical model and its time integrator; it contains all necessary information to execute a model run. Component models subtype ClimaCoupler.AbstractSimulation so that coupling methods can be implmented and dispatch by model type. Component Simulations are collected in a CoupledSimulation along with the coupler itself. This is the full coupled modeling system.

When coupling, component models must implement three methods:

step!: advances the component model by a specified step size

coupler_push!: prepares and puts coupled fields from the model into the coupler via coupler_put! calls

coupler_pull!: gets coupled fields from the coupler via coupler_get! calls and prepares them to be ingested by the model.

These methods hide each components backend implementations (for timestepping or field access) from the coupler, allowing heterogeneity in model backends and the standardization of some coupled timestepping schemes. However, like any other simulation, a CoupledSimulation may also implement a ClimaCoupler.step! method, which in this context specifies the coupling scheme details (e.g. explicit, leap-frog, concurrent, etc...).

ClimaCoupler.AbstractSimulation — Type

abstract type AbstractSimulation

AbstractSimulation

An abstract type representing a model simulation.

Fields:

Contructors:

ClimaCoupler.CoupledSimulation — Type

struct CoupledSimulation{CS, S, CPL, L, C} <: ClimaCoupler.AbstractCoupledSimulation

A collection of coupled component simulations.

Fields:

coupler_solver

The coupled time-stepping scheme

simulations

The component simulations

coupler

The coupler

logger

Diagnostic logger

clock

Clock

Contructors:

CoupledSimulation(
    coupler_solver,
    simulations,
    coupler,
    logger,
    clock
)

defined at /home/runner/work/ClimaCoupler.jl/ClimaCoupler.jl/src/CoupledSimulations/coupled_simulation.jl:26.

ClimaCoupler.run! — Function

run!(sim)

run!(::CoupledSimulation)

A simple outer timestepping loop for coupled system runs.

This will be formalized when the run! functionality for component models is implemented so to have a consistent interface.

ClimaCoupler.step! — Function

step!(sim, dt)

step!(sim, dt)

Advances a simulation by dt.

Note that dt is not necessarily the simulation's timestep length; a simuation could take several shorter steps that total to dt.