Coupled Simulations & Timestepping

ClimaCoupler.jl organizes coupled models and their execution via the CoupledSimulation interface. An implementation of a CoupledSimulation, like any CliMA simulation, must implement a ClimaCoupler.step! method, which in this context specifies the coupling scheme details (e.g. explicit, leap-frog, concurrent, etc...).

Component models being coupled must provide three methods for use in a CoupledSimulation's ClimaCoupler.step! method:

step!: advances the component model in time

coupler_push!: prepares and puts coupled fields from the model into the coupler via coupler_put! calls

coupler_pull!: gets coupled fields from the coupler via coupler_get! calls and prepares them to be ingested by the model.

ClimaCoupler.CoupledSimulation — Type

CoupledSimulation

A subtype of the abstract type AbstractCoupledSimulation representing a model simulation.

ClimaCoupler.run! — Function

run!(::CoupledSimulation)

A simple outer timestepping loop for coupled system runs.

This will be formalized when the run! functionality for component models is implemented so to have a consistent interface.

ClimaCoupler.step! — Function

step!(sim, dt)

Advances a simulation sim by dt.

Note that dt is not necessarily the simulation's timestep length; a simuation could take several shorter steps that total to dt.

ClimaCoupler.coupler_push! — Function

coupler_push!(coupler::CouplerState, model)

Update coupler with fields retrieved from the coupler.

coupler_push! is an adapter function to be implemented for each model component using the coupler. It should send coupling fields via coupler_put! calls and perform any operations on these fields to prepare them for the coupler.

ClimaCoupler.coupler_pull! — Function

coupler_pull!(model, coupler::CouplerState)

Update model with fields retrieved from the coupler.

coupler_pull! is an adapter function to be implemented for each model component using the coupler. It should get coupling fields via coupler_get calls and perform any operations on these fields to prepare them for use in the component model.