Utilities

This module contains functions, objects, and constants used by various modules in the coupler.

Utilities API

Missing docstring.

Missing docstring for ClimaCoupler.Utilities.CouplerSimulation. Check Documenter's build log for details.

ClimaCoupler.Utilities.float_type — Function

float_type(::CoupledSimulation)

Return the floating point type backing T: T can either be an object or a type.

ClimaCoupler.Utilities.swap_space! — Function

swap_space!(field_out::Fields.Field, field_in::Fields.Field)

Remap the values of a field onto a new space.

Arguments

field_in: [Fields.Field] to be remapped to new space.

field_out: [Fields.Field] to remap field_in to.