Utilities

This module contains functions, objects, and constants used by various modules in the coupler.

Utilities API

ClimaCoupler.Utilities.swap_space! — Function

swap_space!(field_out::Fields.Field, field_in::Fields.Field)

Remap the values of a field onto a new space.

Arguments

field_in: [Fields.Field] to be remapped to new space.

field_out: [Fields.Field] to remap field_in to.

ClimaCoupler.Utilities.get_comms_context — Function

get_comms_context(parsed_args)

Sets up the appropriate ClimaComms context for the device the model is to be run on

Arguments

parsed_args: dictionary containing a "device" flag whcih decides which device context is needed

ClimaCoupler.Utilities.get_device — Function

get_device(parsed_args)

Returns the device on which the model is being run

Arguments

parsed_args: dictionary containing a "device" flag which decides which device to run on